Couple bash fixes. rm -rf can cause delete of system root

+++ daily_tests/daily_scl_tests.sh
@@ -21,24 +21,23 @@ TMP_DIR="/tmp/daily_scl_tests"
 RESULT_DIR="${TMP_DIR}/results"
 
 if [[ -d "${TMP_DIR}" ]]; then
-    rm -rf "${TMP_DIR}/"
+    rm -rf "${TMP_DIR:?}/"
 fi
 mkdir -p "${RESULT_DIR}"
 
 function clone_repo() {
     local repo_name=$1; shift
     git clone "https://github.com/sclorg/${repo_name}.git"
-    cd ${repo_name}
+    cd "${repo_name}" || { echo "Repository ${repo_name} does not exist. Skipping." && return 1 ; }
     git submodule update --init
     git submodule update --remote
 }
 
 function iterate_over_all_containers() {
     for repo in ${SCL_CONTAINERS}; do
-        cd ${TMP_DIR}
+        cd ${TMP_DIR} || exit
         local log_name="${TMP_DIR}/${repo}.log"
-        clone_repo "$repo"
-        make test TARGET=centos7 > "${log_name}" 2>&1 || cp "${log_name}" "${RESULT_DIR}/"
+        clone_repo "$repo" && make test TARGET=centos7 > "${log_name}" 2>&1 || cp "${log_name}" "${RESULT_DIR}/"
     done
 }
